user route fixed

+++ src/controllers/user/DeleteUserController.ts
@@ -3,9 +3,12 @@ import DeleteUserService from "../../services/user/DeleteUserService";
 
 class DeleteUserController {
   async handle(req: Request, res: Response) {
-    const { Useruuid } = req.body;
+    const { userId } = req.params;
+    if (!userId) {
+      return res.status(400).json({ error: "User ID is required" });
+    }
     const deleteUserService = new DeleteUserService();
-    const deletedUser = await deleteUserService.execute(Useruuid);
+    const deletedUser = await deleteUserService.execute(userId);
     return res.json(deletedUser);
   }
 }
